Test-first bug hunt (6 fixes): BSGS hash overflow (9% silent miss),
PrevBit doubling, PowerfulNumbers sqrt undershoot, Narayana formula,
QBinomial exponent + q=1 degenerate, Prufer.Rank dead code.
+ ~80 new tests. Suite: 1360/0/0 across 99 projects.

# Conflicts:
#	src/IAFahim.Graph.SCC/TarjanScc.cs

+++ src/IAFahim.Math.NT/MoebiusPrefix.cs
@@ -35,7 +35,7 @@ namespace IAFahim.Math.NT
                     primes[pCount++] = i;
                     mu[i] = -1;
                 }
-                for (int j = 0; j < pCount && i * primes[j] <= n; j++)
+                for (int j = 0; j < pCount && (long)i * primes[j] <= n; j++)
                 {
                     int p = primes[j];
                     isPrime[i * p] = false;
